Added majority of ocpp smart charging profile
- Added SetChargingProfile
- Added ClearChargingProfile
- Added selecting of active charging profile
- Added updating of charging profiles based on transaction information,
  profile validity range and duration, ocpp messages changing profiles.
- Added changing charging variables (min/max current) based on current
  charging period.
- Changed Meter value for current.offered.

There is an issue rdegarding file handeling that conflicts with
offlineSession and offlineLog. This issue is due to the way
esp-idf/FatFs handles translation of volume to partition. esp-idf commit
1bad4f18ed3adf7397b9c53967f8a8d9e7cc959c and 63b96d7612ae41db4db8dd5cf93f83c31ed29336
should be considered. Alternatively FF_VOLUMES should be reverted to 2
and each partition should only be mounted once. Both alternatives may be
used together.