Attempt to fix CRC mismatch for stop transaction when stop related meter value was present. See https://app.clickup.com/t/86933j969 for more info.
The changes assumes the issue is related to a
race condition where next transaction message is
sent after meter value was written, but before
stop transaction was written. It is hard to test
if the changes fixes the issue.

One thing to note is that meter value and stop
transaction do not share a crc value. If a crash
prevents the stop transaction, then there could
be a mismatch between the timestamps.